refactor(DC-704): migrate IdentifierHasher startup test onto StartupValidationTestBase
